API Reference

Dimensionless numbers

PlasmaFormulary.plasma_beta — Function

plasma_beta(T, n, B)

Compute the plamsa beta (β), the ratio of thermal pressure to magnetic pressure.

Arguments

T: The temperature of the plasma.

n: The particle density of the plasma.

B: The magnetic field in the plasma.

Example

julia> plasma_beta(1e6u"K", 1e19u"m^-3", 0.2u"T")
0.008674873511172188

References

PlasmaPy Documentation Wikipedia

Otherwise undocumented functions

This section will be removed once the documentation is complete.

PlasmaFormulary.plasma_frequency — Method

plasma_frequency(n, Z, mass_numb)

Ion plasma frequency.

PlasmaFormulary.plasma_frequency — Method

plasma_frequency(n, [q, mass])

Particle plasma frequency, often this is the cold electrons plasma frequency.

References:

https://en.wikipedia.org/wiki/Plasma_oscillation

PlasmaFormulary.gyroradius — Method

Calculate the radius of circular motion for a charged particle in a uniform magnetic field

References: PlasmaPy API Documentation

Examples

julia> gyroradius(0.2u"T", Unitful.me, Unitful.q, 1e6u"K")
0.00015651672339994665 m

PlasmaFormulary.inertial_length — Method

The inertial length is the characteristic length scale for a particle to be accelerated in a plasma. The Hall effect becomes important on length scales shorter than the ion inertial length.

References: PlasmaPy API Documentation

PlasmaFormulary.magnetic_pressure — Method

magnetic_pressure(B)

Calculate the magnetic pressure.

PlasmaFormulary.thermal_pressure — Method

thermal_pressure(T, n)

Calculate the thermal pressure for a Maxwellian distribution.

Arguments

T: The particle temperature or energy.

n: The particle number density.

PlasmaFormulary.alfven_velocity — Method

The typical propagation speed of magnetic disturbances in a quasineutral plasma.

References: PlasmaPy API Documentation

PlasmaFormulary.thermal_velocity_coefficients — Method

thermal_velocity_coefficients(method::ThermalVelocityMethod, ndim::Int)

Get the thermal speed coefficient corresponding to the desired thermal speed definition.

Arguments

method::ThermalVelocityMethod: Method to be used for calculating the thermal speed. Valid values are MostProbable(), RMS(), MeanMagnitude(), and NRL().

ndim::Val{Int}: Dimensionality (1D, 2D, 3D) of space in which to calculate thermal speed. Valid values are Val(1), Val(2), or Val{3}.